% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSmoothGradient.R
\name{addSmoothGradient}
\alias{addSmoothGradient}
\title{Add smoothed gradient terms to data}
\usage{
addSmoothGradient(
  data,
  Time.name = NULL,
  Time.unit = "auto",
  spatialCovs,
  weights = c(1/2, 1/8, 1/8, 1/8, 1/8),
  scale = NULL
)
}
\arguments{
\item{data}{A \code{\link{momentuHMMData}} object (as returned by \code{\link{prepData}} or \code{\link{simData}}).}

\item{Time.name}{Character string indicating name of the time column in \code{data} (for continuous-time HMMs). Default: 'NULL' (discrete time).}

\item{Time.unit}{Character string indicating units for time difference between observations (e.g. 'auto', 'secs', 'mins', 'hours', 'days', 'weeks'). Ignored unless \code{data[[Time.name]]} is of class \code{\link[base]{date-time}} or \code{\link[base]{date}}. Default: 'auto'.}

\item{spatialCovs}{List of \code{\link[raster]{raster}} objects for spatio-temporally referenced covariates. Covariates specified by \code{spatialCovs} are extracted from the raster 
layer(s) based on the location data (and the z values for a raster \code{\link[raster]{stack}} 
or \code{\link[raster]{brick}}) for each time step.  If an element of \code{spatialCovs} is a raster \code{\link[raster]{stack}} or \code{\link[raster]{brick}}, 
then z values must be set using \code{raster::setZ} and \code{data} must include column(s) of the corresponding z value(s) for each observation (e.g., 'time').}

\item{weights}{Numeric vector indicating the weight for the points in a smoothed version of the gradient, where \code{length(weights)}\eqn{\in \{5,9\}} defines the number of points and \code{sum(weights)=1}. Default: \code{c(1/2,1/8,1/8,1/8,1/8)}, where the first weight (1/2) corresponds to the gradient at the observed location, with the remaining weights (1/8, 1/8, 1/8, 1/8) corresponding to clockwise points around the observed location (i.e. northeast, southeast, southwest, northwest). The smoothed gradients are returned with ``\code{.xs}'' (easting gradient) and ``\code{.ys}'' (northing gradient) suffixes added to the names of \code{spatialCovs}.}

\item{scale}{Numeric scalar indicating the distance from the observed location to each smoothing point (before adjusting by time step for continuous-time HMMs). Default: \code{NULL}, in which case the square root of the mean squared step lengths (divided by the time step for continuous-time HMMs) across all individuals is used, that is, \code{scale}\eqn{=\sqrt{0.5*mean(step^2/dt)}}, where \eqn{step} is step length and \eqn{dt} is the time step (note \eqn{dt=1} for discrete-time HMMs). The smoothing points are placed at a distance of \code{scale}\eqn{\sqrt{2 dt}} from the observed locations.}
}
\value{
A \code{\link{momentuHMM}} or \code{\link{momentuHierHMM}} object, with the smooth gradients added for each spatial covariate.
}
\description{
Adds smoothed gradients to \code{\link{momentuHMMData}} or \code{\link{momentuHierHMMData}} objects, which can help reduce bias in gradient-based habitat selection coefficients
}
\references{
Blackwell, P. G. and J. Matthiopoulos. 2024. Joint inference for telemetry and spatial survey data. Ecology
}
