% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angleCI.R
\name{angleCI}
\alias{angleCI}
\title{Confidence intervals for angle parameters}
\usage{
angleCI(m, alpha, nbSims = 10^6)
}
\arguments{
\item{m}{A \code{momentuHMM} object}

\item{alpha}{Range of the confidence intervals. Default: 0.95 (i.e. 95\% CIs).}

\item{nbSims}{Number of simulations. Default: 10^6.}
}
\value{
A list of the following objects:
\item{lower}{Lower bound of the confidence interval for the parameters of the angle distribution}
\item{upper}{Upper bound of the confidence interval for the parameters of the angle distribution}
}
\description{
Simulation-based computation of confidence intervals for the parameters of the angle distribution.
Used in \code{\link{CI_real}}.
}

