% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHierPar0.R
\name{checkHierPar0}
\alias{checkHierPar0}
\title{Check parameter length and order for a \code{\link{fitHierHMM}} (or \code{\link{MIfitHierHMM}}) model}
\usage{
checkHierPar0(data, hierStates, hierDist, Par0 = NULL, beta0 = NULL,
  delta0 = NULL, estAngleMean = NULL, circularAngleMean = NULL,
  hierFormula = NULL, formulaDelta = ~1, mixtures = 1, formulaPi = NULL,
  DM = NULL, userBounds = NULL, workBounds = NULL, betaCons = NULL,
  fixPar = NULL)
}
\arguments{
\item{data}{\code{\link{momentuHierHMMData}} object or a data frame containing the data stream and covariate values}

\item{hierStates}{A hierarchical model structure \code{\link[data.tree]{Node}} for the states ('state').  See details.}

\item{hierDist}{A hierarchical data structure \code{\link[data.tree]{Node}} for the data streams ('dist'). Currently
supported distributions are 'bern', 'beta', 'exp', 'gamma', 'lnorm', 'norm', 'mvnorm2' (bivariate normal distribution), 'mvnorm3' (trivariate normal distribution),
'pois', 'rw_norm' (normal random walk), 'rw_mvnorm2' (bivariate normal random walk), 'rw_mvnorm3' (trivariate normal random walk), 'vm', 'vmConsensus', 'weibull', and 'wrpcauchy'. See details.}

\item{Par0}{Optional named list containing vectors of state-dependent probability distribution parameters for 
each data stream specified in \code{dist}.  If \code{Par0} is not provided, then ordered parameter indices are returned.}

\item{beta0}{Optional matrix of regression coefficients for the transition probabilities. If \code{beta0} is not provided, then ordered parameter indices are returned.}

\item{delta0}{Optional values or regression coefficients for the initial distribution of the HMM. If \code{delta0} is not provided, then ordered parameter indices are returned.}

\item{estAngleMean}{An optional named list indicating whether or not to estimate the angle mean for data streams with angular 
distributions ('vm' and 'wrpcauchy').}

\item{circularAngleMean}{An optional named list indicating whether to use circular-linear or circular-circular
regression on the mean of circular distributions ('vm' and 'wrpcauchy') for turning angles.}

\item{hierFormula}{A hierarchical formula structure for the transition probability covariates for each level of the hierarchy ('formula'). Default: \code{NULL} (no covariate effects). In addition to allowing standard functions in R formulas
(e.g., \code{cos(cov)}, \code{cov1*cov2}, \code{I(cov^2)}), special functions include \code{cosinor(cov,period)} for modeling cyclical patterns, spline functions 
(\code{\link[splines]{bs}}, \code{\link[splines]{ns}}, \code{\link[splines2]{bSpline}}, \code{\link[splines2]{cSpline}}, \code{\link[splines2]{iSpline}}, and \code{\link[splines2]{mSpline}}), 
and state- or parameter-specific formulas (see details).
Any formula terms that are not state- or parameter-specific are included on all of the transition probabilities within a given level of the hierarchy.}

\item{formulaDelta}{Regression formula for the initial distribution.}

\item{mixtures}{Number of mixtures for the state transition probabilities.}

\item{formulaPi}{Regression formula for the mixture distribution probabilities. 
Note that only the covariate values from the first row for each individual ID in \code{data} are used (i.e. time-varying covariates cannot be used for the mixture probabilties).}

\item{DM}{An optional named list indicating the design matrices to be used for the probability distribution parameters of each data 
stream.}

\item{userBounds}{An optional named list of 2-column matrices specifying bounds on the natural (i.e, real) scale of the probability 
distribution parameters for each data stream.}

\item{workBounds}{An optional named list of 2-column matrices specifying bounds on the working scale of the probability distribution, transition probability, and initial distribution parameters.}

\item{betaCons}{Matrix of the same dimension as \code{beta0} composed of integers identifying any equality constraints among the t.p.m. parameters.}

\item{fixPar}{An optional list of vectors indicating parameters which are assumed known prior to fitting the model.}
}
\description{
Prints parameters with labels based on \code{DM}, \code{hierFormula}, and/or \code{formulaDelta}.  See \code{\link{fitHierHMM}} for 
further argument details.
}
\seealso{
\code{\link{fitHierHMM}}, \code{\link{MIfitHierHMM}}
}
