% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPar0.R
\name{checkPar0}
\alias{checkPar0}
\title{Check parameter length and order of a \code{\link{fitHMM}} (or \code{\link{MIfitHMM}}) model}
\usage{
checkPar0(data, nbStates, dist, Par0, beta0 = NULL, delta0 = NULL,
  estAngleMean = NULL, circularAngleMean = NULL, formula = ~1,
  formulaDelta = ~1, stationary = FALSE, DM = NULL, cons = NULL,
  userBounds = NULL, workcons = NULL, stateNames = NULL, fixPar = NULL)
}
\arguments{
\item{data}{An object \code{momentuHMMData}.}

\item{nbStates}{Number of states of the HMM.}

\item{dist}{A named list indicating the probability distributions of the data streams.}

\item{Par0}{A named list containing vectors of initial state-dependent probability distribution parameters for 
each data stream specified in \code{dist}.}

\item{beta0}{Initial matrix of regression coefficients for the transition probabilities.}

\item{delta0}{Initial value for the initial distribution of the HMM.}

\item{estAngleMean}{An optional named list indicating whether or not to estimate the angle mean for data streams with angular 
distributions ('vm' and 'wrpcauchy').}

\item{circularAngleMean}{An optional named list indicating whether to use circular-linear (FALSE) or circular-circular (TRUE) 
regression on the mean of circular distributions ('vm' and 'wrpcauchy') for turning angles.}

\item{formula}{Regression formula for the transition probability covariates.}

\item{formulaDelta}{Regression formula for the initial distribution.}

\item{stationary}{\code{FALSE} if there are covariates in \code{formula} or \code{formulaDelta}. If \code{TRUE}, the initial distribution is considered
equal to the stationary distribution. Default: \code{FALSE}.}

\item{DM}{An optional named list indicating the design matrices to be used for the probability distribution parameters of each data 
stream.}

\item{cons}{An optional named list of vectors specifying a power to raise parameters corresponding to each column of the design matrix 
for each data stream.}

\item{userBounds}{An optional named list of 2-column matrices specifying bounds on the natural (i.e, real) scale of the probability 
distribution parameters for each data stream.}

\item{workcons}{An optional named list of vectors specifying constants to add to the regression coefficients on the working scale for 
each data stream.}

\item{stateNames}{Optional character vector of length nbStates indicating state names.}

\item{fixPar}{An optional list of vectors indicating parameters which are assumed known prior to fitting the model.}
}
\description{
Verify initial parameters (\code{Par0}, \code{beta0}, and \code{delta0}) are of correct length and print the parameters with labels based on \code{DM}, \code{formula}, and/or \code{formulaDelta}.  See \code{\link{fitHMM}} for 
further argument details.
}
\seealso{
\code{\link{fitHMM}}, \code{\link{MIfitHMM}}
}
