% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_beta.R
\name{CI_beta}
\alias{CI_beta}
\title{Confidence intervals}
\usage{
CI_beta(m, alpha = 0.95)
}
\arguments{
\item{m}{A \code{momentuHMM} object}

\item{alpha}{Range of the confidence intervals. Default: 0.95 (i.e. 95\% CIs).}
}
\value{
A list of the following objects:
\item{stepPar}{Standard errors and confidence intervals for the working parameters of the step lengths distribution ('gammabeta' or 'weibullbeta' depending on stepDist)}
\item{anglePar}{Standard errors and confidence intervals for the working parameters of the turning angles distribution ('concbeta' or 'sdbeta' depending on angleDist)}
\item{omegaPar}{Standard errors and confidence intervals for the working parameters of the omega distribution ('omegabeta')}
\item{dryPar}{Standard errors and confidence intervals for the working parameters of the dry distribution ('drybeta')}
\item{divePar}{Standard errors and confidence intervals for the working parameters of the dive distribution ('divebeta')}
\item{icePar}{Standard errors and confidence intervals for the working parameters of the ice distribution ('icebeta')}
\item{landPar}{Standard errors and confidence intervals for the working parameters of the land distribution ('landbeta')}
\item{betaPar}{Standard errors and confidence intervals for the working parameters of the transition probabilities}
}
\description{
Computes the standard errors and confidence intervals on the beta (i.e., working) scale of the step length and turning angle parameters,
as well as for the transition probabilities regression parameters. Working scale depends on the real (i.e., natural) scale of the parameters. For 
non-circular distributions or for circular distributions with \code{estAngleMean}=FALSE:
}
\details{
1) if both lower and upper bounds are finite then logit is the working scale;
2) if lower bound is finite and upper bound is infinite then log is the working scale.

For circular distributions with \code{estAngleMean}=TRUE and no constraints imposed by a design matrix (DM) or bounds (userBounds), then the working parameters 
are complex functions of both the angle mean and concentrations/sd natural parameters (in this case, it's probably best just to focus on the real parameter
estimates!).  However, if constraints are imposed by DM or userBounds on circular distribution parameters with \code{estAngleMean}=TRUE:

1) if the natural bounds are (-pi,pi] then tangent is the working scale, otherwise if both lower and upper bounds are finite then logit is the working scale;
2) if lower bound is finite and upper bound is infinite then log is the working scale.
}
\examples{
# m is a momentuHMM object (as returned by fitHMM), automatically loaded with the package
m <- example$m

CI_beta(m)

}
