% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_real.R
\name{CI_real}
\alias{CI_real}
\title{Confidence intervals}
\usage{
CI_real(m, alpha = 0.95, nbSims = 10^6)
}
\arguments{
\item{m}{A \code{momentuHMM} object}

\item{alpha}{Range of the confidence intervals. Default: 0.95 (i.e. 95\% CIs).}

\item{nbSims}{Number of simulations in the computation of the CIs for the angle parameters.
Default: 10^6.}
}
\value{
A list of the following objects:
\item{stepPar}{Standard errors and confidence intervals for the parameters of the step lengths distribution}
\item{anglePar}{Standard errors and confidence intervals for the parameters of the turning angles distribution}
\item{omegaPar}{Standard errors and confidence intervals for the parameters of the omega distribution}
\item{dryPar}{Standard errors and confidence intervals for the parameters of the dry distribution}
\item{divePar}{Standard errors and confidence intervals for the parameters of the dive distribution}
\item{icePar}{Standard errors and confidence intervals for the parameters of the ice distribution}
\item{landPar}{Standard errors and confidence intervals for the parameters of the land distribution}
\item{gamma}{Standard errors and confidence intervals for the transition probabilities. If covariates are included in TPM formula, then the covariate values for the first observation in the data are used.}
}
\description{
Computes the standard errors and confidence intervals on the real (i.e., natural) scale of the step length and turning angle parameters,
as well as for the transition probabilities regression parameters.
}
\examples{
# m is a momentuHMM object (as returned by fitHMM), automatically loaded with the package
m <- example$m

CI_real(m)

}
