% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circAngles.R
\name{circAngles}
\alias{circAngles}
\title{Convert bearings (relative to true north) to turning angle covariates suitable for circular-circular regression}
\usage{
circAngles(refAngle, data, coordNames = c("x", "y"))
}
\arguments{
\item{refAngle}{Numeric vector of bearings relative to true north}

\item{data}{data frame containing fields for the x- and y-coordinates (identified by \code{coordNames}) and 'ID' (if more than one individual)}

\item{coordNames}{Names of the columns of coordinates in \code{data}. Default: \code{c("x","y")}.}
}
\value{
A vector of turning angles between the movement direction at time step t-1 and \code{refAngle} at time t
}
\description{
This function can be used to convert angular covariates measured relative to true north (e.g., ocean currents, wind direction) to turning angle
covariates sutiable for circular-circular regression in \code{\link{fitHMM}} or \code{\link{MIfitHMM}}.
}
\examples{
# extract data from momentuHMM example
data<-example$data

# generate fake bearings
refAngle<-2*atan(rnorm(nrow(data)))

# add turning angle covariate to data
data$cov3 <- circAngles(refAngle=refAngle,data=data)

}
