% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentuHMM_utils.R
\name{collapseRaster}
\alias{collapseRaster}
\title{Transform a raster into a (x,y,z) list}
\usage{
collapseRaster(rast)
}
\arguments{
\item{rast}{\code{\link[raster]{raster}} layer object for spatially referenced covariates.}
}
\value{
List of three elements: x (grid of x values), y (grid of y values), and z (matrix of values).
}
\description{
Transform a raster into a (x,y,z) list
}
