% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crwData.R
\name{crwData}
\alias{crwData}
\title{Constructor of \code{crwData} objects}
\usage{
crwData(m)
}
\arguments{
\item{m}{A list of attributes of crawl output: \code{crwFits} (a list of crwFit objects) and \code{crwPredict} (a crwPredict object)}
}
\value{
An object \code{crwData}.
}
\description{
Constructor of \code{crwData} objects
}
\seealso{
\code{\link{crawlWrap}}, \code{\link{MIfitHMM}}
}
