% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crwHierData.R
\name{crwHierData}
\alias{crwHierData}
\title{Constructor of \code{crwHierData} objects}
\usage{
crwHierData(m)
}
\arguments{
\item{m}{A list of attributes of crawl output: \code{crwFits} (a list of crwFit objects) and \code{crwPredict} (a crwPredict object)}
}
\value{
An object \code{crwHierData}.
}
\description{
Constructor of \code{crwHierData} objects
}
\seealso{
\code{\link{crawlWrap}}, \code{\link{MIfitHMM}}
}
