% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crwHierSim.R
\name{crwHierSim}
\alias{crwHierSim}
\title{Constructor of \code{crwHierSim} objects}
\usage{
crwHierSim(m)
}
\arguments{
\item{m}{A list of attributes required for multiple imputation data generated from a \code{\link{crwHierData}} object using \code{\link{MIfitHMM}}: \code{miData} (a list of 
\code{\link{momentuHMMData}} objects), and \code{crwSimulator} (a list of \code{\link[crawl]{crwSimulator}} objects).

\code{crwHierSim} objects are returned by \code{\link{MIfitHMM}} when argument \code{miData} is a \code{\link{crwHierData}} object and argument \code{fit=FALSE}.}
}
\value{
An object \code{crwHierSim}.
}
\description{
Constructor of \code{crwHierSim} objects
}
