% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crwSim.R
\name{crwSim}
\alias{crwSim}
\title{Constructor of \code{crwSim} objects}
\usage{
crwSim(m)
}
\arguments{
\item{m}{A list of attributes required for multiple imputation data generated from a \code{\link{crwData}} object using \code{\link{MIfitHMM}}: \code{miData} (a list of 
\code{\link{momentuHMMData}} objects), and \code{crwSimulator} (a list of \code{\link[crawl]{crwSimulator}} objects).

\code{crwSim} objects are returned by \code{\link{MIfitHMM}} when argument \code{miData} is a \code{\link{crwData}} object and argument \code{fit=FALSE}.}
}
\value{
An object \code{crwSim}.
}
\description{
Constructor of \code{crwSim} objects
}
