% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdsUD.R
\name{ctdsUD}
\alias{ctdsUD}
\title{Calculate and plot the (state-dependent) stationary (or utilization) distribution from a \code{ctds} fitted model object}
\usage{
ctdsUD(
  ctds,
  spatialCovs,
  spatialCovs.grad,
  zero.idx = integer(),
  method = "lu",
  maxiter,
  start,
  tol
)
}
\arguments{
\item{ctds}{A \code{ctds} object returned by \code{\link{fitCTHMM}}}

\item{spatialCovs}{List of \code{\link[raster]{raster}} objects for spatially referenced covariates. Raster \code{\link[raster]{stack}}s or \code{\link[raster]{brick}}s are not currently allowed.}

\item{spatialCovs.grad}{List of \code{\link[raster]{raster}} objects for spatially referenced covariates, where a directional gradient is to be calculated internally using \code{\link[ctmcmove]{rast.grad}}. Raster \code{\link[raster]{stack}}s or \code{\link[raster]{brick}}s are not currently allowed.}

\item{zero.idx}{Integer vector of the indices of raster cells that are not passable and should be excluded. These are cells where movement should be impossible. Default is zero.idx=integer().}

\item{method}{Either "lu" (default) or "limit". See \code{\link[ctmcmove]{get.UD}} details for a description of the two methods.}

\item{maxiter}{Total number of iterations for limit method if tolerance not reached first. Defaults to 100. Ignored for method='lu'.}

\item{start}{A value for the starting distribution for the 'limit' method. Defaults to 1/num. cells. Ignored for method='lu'.}

\item{tol}{Value used to assess convergence for limit method. If max(abs(pi1-pi0))<tol, limit method has converged. Defaults to sqrt(.Machine$double.eps).}
}
\value{
A list of \code{\link[raster]{raster}} objects.
}
\description{
This is essentially a wrapper function for \code{\link[ctmcmove]{get.rate.matrix}} and \code{\link[ctmcmove]{get.UD}} that has been extended for \code{\link{momentuHMM}}.
}
