% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbern_rcpp}
\alias{dbern_rcpp}
\title{Bernoulli density function}
\usage{
dbern_rcpp(x, prob, foo)
}
\arguments{
\item{x}{Vector of quantiles}

\item{prob}{success probability}

\item{foo}{Unused (for compatibility with template)}
}
\value{
Vector of densities
}
\description{
Probability density function of the Bernoulli distribution (written in C++)
}
