% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbeta_rcpp}
\alias{dbeta_rcpp}
\title{Probability density function of the beta distribution (written in C++)}
\usage{
dbeta_rcpp(x, shape1, shape2)
}
\arguments{
\item{x}{Vector of quantiles}

\item{shape1}{Shape1}

\item{shape2}{Shape2}
}
\value{
Vector of densities
}
\description{
Probability density function of the beta distribution (written in C++)
}
