% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dcat_rcpp}
\alias{dcat_rcpp}
\title{Categorical density function}
\usage{
dcat_rcpp(x, prob, foo)
}
\arguments{
\item{x}{Vector of quantiles}

\item{prob}{success probability}

\item{foo}{Unused (for compatibility with template)}
}
\value{
Vector of densities
}
\description{
Probability density function of the categorical distribution (written in C++)
}
