% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dcrwrice_rcpp}
\alias{dcrwrice_rcpp}
\title{Correlated random walk Rice distribution}
\usage{
dcrwrice_rcpp(x, beta, sigma)
}
\arguments{
\item{x}{numeric data vector of length \code{n + n} where the first \code{n} entries correspond to the step lengths and the next \code{n} entries to the corresponding previous step lengths (the first of which is NA and ignored).}

\item{beta}{correlation parameter}

\item{sigma}{speed parameter}
}
\value{
Vector of densities
}
\description{
Probability density function of Rice distribution under the CTCRW model of Johnson et al. (2008) (written in C++)
}
