% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dcrwvm_rcpp}
\alias{dcrwvm_rcpp}
\title{Correlated random walk von Mises density function}
\usage{
dcrwvm_rcpp(x, beta, sigma)
}
\arguments{
\item{x}{numeric data vector of length \code{n + n + n} where the first \code{n} entries correspond to angles (von Mises distribution), the next \code{n} entries to the corresponding step lengths, and the last \code{n} entries to the corresponding previous step lengths.}

\item{beta}{correlation parameter}

\item{sigma}{speed parameter}
}
\value{
Vector of densities
}
\description{
Probability density function of the Von Mises distribution under the CTCRW model of Johnson et al. (2008) (written in C++)
}
