% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dexp_rcpp}
\alias{dexp_rcpp}
\title{Exponential density function}
\usage{
dexp_rcpp(x, rate, foo)
}
\arguments{
\item{x}{Vector of quantiles}

\item{rate}{Rate}

\item{foo}{Unused (for compatibility with template)}
}
\value{
Vector of densities
}
\description{
Probability density function of the exponential distribution (written in C++)
}
