% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dgamma_rcpp}
\alias{dgamma_rcpp}
\title{Gamma density function}
\usage{
dgamma_rcpp(x, mu, sigma)
}
\arguments{
\item{x}{Vector of quantiles}

\item{mu}{Mean}

\item{sigma}{Standard deviation}
}
\value{
Vector of densities
}
\description{
Probability density function of the gamma distribution (written in C++)
}
