% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distAngle.R
\name{distAngle}
\alias{distAngle}
\title{Calculate distance between points y and z and turning angle between points x, y, and z}
\usage{
distAngle(x, y, z)
}
\arguments{
\item{x}{location 1}

\item{y}{location 2}

\item{z}{location 3}
}
\value{
2-vector with first element the distance between y and z and second element the turning angle
between (x,y) and (y,z).
}
\description{
Calculate distance between points y and z and turning angle between points x, y, and z
}
\details{
Used in \code{\link{prepData}} and \code{\link{simData}} to get distance and turning angle covariates 
between locations (x1,x2), (y1,y2) and activity center (z1,z2).
}
