% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlnorm_rcpp}
\alias{dlnorm_rcpp}
\title{Log-normal density function}
\usage{
dlnorm_rcpp(x, meanlog, sdlog)
}
\arguments{
\item{x}{Vector of quantiles}

\item{meanlog}{Mean of the distribution on the log-scale}

\item{sdlog}{Standard deviation of the distribution on the log-scale}
}
\value{
Vector of densities
}
\description{
Probability density function of the log-normal distribution (written in C++)
}
