% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlogis_rcpp}
\alias{dlogis_rcpp}
\title{logistic density function}
\usage{
dlogis_rcpp(x, location, scale)
}
\arguments{
\item{x}{Vector of quantiles}

\item{location}{mean of the distribution}

\item{scale}{Dispersion parameter}
}
\value{
Vector of densities
}
\description{
Probability density function of the logistic distribution (written in C++)
}
