% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnbinom_rcpp}
\alias{dnbinom_rcpp}
\title{negative binomial density function}
\usage{
dnbinom_rcpp(x, mu, size)
}
\arguments{
\item{x}{Vector of quantiles}

\item{mu}{Mean of the distribution}

\item{size}{Dispersion parameter}
}
\value{
Vector of densities
}
\description{
Probability density function of the negative binomial distribution (written in C++)
}
