% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnorm_rcpp}
\alias{dnorm_rcpp}
\title{Normal density function}
\usage{
dnorm_rcpp(x, mean, sd)
}
\arguments{
\item{x}{Vector of quantiles}

\item{mean}{Mean of the distribution}

\item{sd}{Standard deviation of the distribution}
}
\value{
Vector of densities
}
\description{
Probability density function of the normal distribution (written in C++)
}
