% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dpois_rcpp}
\alias{dpois_rcpp}
\title{Poisson density function}
\usage{
dpois_rcpp(x, rate, foo)
}
\arguments{
\item{x}{Vector of quantiles}

\item{rate}{Rate}

\item{foo}{Unused (for compatibility with template)}
}
\value{
Vector of densities
}
\description{
Probability density function of the Poisson distribution (written in C++)
}
