% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dt_rcpp}
\alias{dt_rcpp}
\title{student t density function}
\usage{
dt_rcpp(x, df, ncp)
}
\arguments{
\item{x}{Vector of quantiles}

\item{df}{degrees of freedom}

\item{ncp}{non-centrality parameter}
}
\value{
Vector of densities
}
\description{
Probability density function of non-central student t (written in C++)
}
