% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dvm_rcpp}
\alias{dvm_rcpp}
\title{Von Mises density function}
\usage{
dvm_rcpp(x, mu, kappa)
}
\arguments{
\item{x}{Vector of quantiles}

\item{mu}{Mean}

\item{kappa}{Concentration}
}
\value{
Vector of densities
}
\description{
Probability density function of the Von Mises distribution, defined as a function
of the modified Bessel function of order 0 (written in C++)
}
