% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dweibull_rcpp}
\alias{dweibull_rcpp}
\title{Weibull density function}
\usage{
dweibull_rcpp(x, shape, scale)
}
\arguments{
\item{x}{Vector of quantiles}

\item{shape}{Shape}

\item{scale}{Scale}
}
\value{
Vector of densities
}
\description{
Probability density function of the Weibull distribution (written in C++)
}
