% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dwrpcauchy_rcpp}
\alias{dwrpcauchy_rcpp}
\title{Wrapped Cauchy density function}
\usage{
dwrpcauchy_rcpp(x, mu, rho)
}
\arguments{
\item{x}{Vector of quantiles}

\item{mu}{Mean}

\item{rho}{Concentration}
}
\value{
Vector of densities
}
\description{
Probability density function of the wrapped Cauchy distribution (written in C++)
}
