% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exGen.R
\docType{data}
\name{example}
\alias{example}
\title{Example dataset}
\usage{
example
}
\description{
This data is generated by the function \code{\link{exGen}}, and used in the examples and tests of
other functions to keep them as short as possible.
}
\details{
It is a list of the following objects:
\itemize{
\item \code{data} A \code{\link{momentuHMMData}} object
\item \code{m} A \code{\link{momentuHMM}} object
\item \code{simPar} The parameters used to simulate \code{data}
\item \code{par0} The initial parameters in the optimization to fit \code{m}
}
}
