% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCovNames.R
\name{getCovNames}
\alias{getCovNames}
\title{Get names of any covariates used in probability distribution parameters}
\usage{
getCovNames(m, p, distname)
}
\arguments{
\item{m}{\code{\link{momentuHMM}} object}

\item{p}{list returned by \code{\link{parDef}}}

\item{distname}{Name of the data stream}
}
\value{
A list of:
\item{DMterms}{Names of all covariates included in the design matrix for the data stream}
\item{DMpartems}{A list of the names of all covariates for each of the probability distribution parameters}
}
\description{
Get names of any covariates used in probability distribution parameters
}
