% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getDM_rcpp}
\alias{getDM_rcpp}
\title{Get design matrix}
\usage{
getDM_rcpp(X, covs, DM, nr, nc, cov, nbObs)
}
\arguments{
\item{X}{full design matrix}

\item{covs}{matrix of covariates}

\item{DM}{pseudo design matrix}

\item{nr}{number of rows in design matrix}

\item{nc}{number of column in design matrix}

\item{cov}{covariate names}

\item{nbObs}{number of observations}
}
\value{
full design matrix (X)
}
\description{
Loop for creating full design matrix (X) from pseudo-design matrix (DM). Written in C++. Used in \code{getDM}.
}
