% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentuHMM_utils.R
\name{getGradients}
\alias{getGradients}
\title{Calculate gradient of spatial covariates using bilinear interpolation}
\usage{
getGradients(data, spatialCovs, coordNames = c("x", "y"))
}
\arguments{
\item{data}{Data frame of data streams. At a minimum, it must contain fields matching \code{coordNames}}

\item{spatialCovs}{List of \code{\link[raster]{raster}} layer objects for spatially referenced covariates. Covariates specified by \code{spatialCovs} are extracted from the raster 
layer(s) based on the location data for each time step.}

\item{coordNames}{Names of the coordinates in \code{data}. Default: \code{c("x","y")}.}
}
\value{
The gradients are appended to \code{data} with ``\code{.x}'' (easting gradient) and ``\code{.y}'' (northing gradient) suffixes added to the names of \code{spatialCovs}. For example, if \code{cov1} is the name of a spatial covariate, then the returned \code{data} object will include the fields ``\code{cov1.x}'' and ``\code{cov1.y}''.
}
\description{
Calculate gradient of spatial covariates using bilinear interpolation
}
