% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPar.R
\name{getPar}
\alias{getPar}
\title{Get starting values (Par) from momentuHMM object returned by fitHMM}
\usage{
getPar(m)
}
\arguments{
\item{m}{A \code{momentuHMM} object.}
}
\value{
A list of parameter values that can be used as starting values (Par) in fitHMM
}
\description{
Get starting values (Par) from momentuHMM object returned by fitHMM
}
\examples{
# m is a momentuHMM object (as returned by fitHMM), automatically loaded with the package
m <- example$m
Par <- getPar(m)

}

