% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParDM.R
\name{getParDM}
\alias{getParDM}
\title{Get starting values (Par) on working scale based on design matrix}
\usage{
getParDM(data, nbStates, dist, Par, estAngleMean = NULL,
  circularAngleMean = NULL, DM = NULL, cons = NULL, userBounds = NULL,
  workcons = NULL, stateNames = NULL)
}
\value{
A list of parameter values (Par) that can be used as starting values in \code{\link{fitHMM}} or \code{\link{MIfitHMM}}
}
\description{
Get starting values (Par) on working scale based on design matrix
}
\examples{
# m is a moveHMMext object (as returned by fitHMM), automatically loaded with the package
m <- example$m
Par <- getPar(m)

}

