% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrProbs.R
\name{getTrProbs}
\alias{getTrProbs}
\alias{getTrProbs.default}
\alias{getTrProbs.hierarchical}
\title{Transition probability matrix}
\usage{
getTrProbs(data, ...)

\method{getTrProbs}{default}(data, nbStates, beta, workBounds = NULL,
  formula = ~1, mixtures = 1, betaRef = NULL, stateNames = NULL,
  ...)

\method{getTrProbs}{hierarchical}(data, hierStates, hierBeta,
  workBounds = NULL, hierFormula = NULL, mixtures = 1, hierDist, ...)
}
\arguments{
\item{data}{\code{\link{momentuHMM}} object, \code{\link{momentuHierHMM}} object, \code{\link{momentuHMMData}} object, \code{\link{momentuHierHMMData}} object, or a data frame containing the covariate values. 

If a data frame is provided, then either \code{nbStates} must be specified (for a regular HMM) or \code{hierStates} and \code{hierDist}
must be specified (for a hierarchical HMM).}

\item{...}{further arguments passed to or from other methods; ignored if \code{data} is a \code{\link{momentuHMM}} or \code{\link{momentuHierHMM}} object}

\item{nbStates}{Number of states. Ignored unless \code{data} is a data frame.}

\item{beta}{Matrix of regression coefficients for the transition probabilities}

\item{workBounds}{An optional named list of 2-column matrices specifying bounds on the working scale of the transition probability parameters ('beta' and, for recharge models, 'g0' and 'theta'). \code{workBounds$beta} must be a k x 2 matrix, where k=\code{length(beta)}.
The first column pertains to the lower bound and the second column the upper bound. Ignored unless \code{data} is a data frame.}

\item{formula}{Regression formula for the transition probability covariates. Ignored unless \code{data} is a data frame.}

\item{mixtures}{Number of mixtures for the state transition probabilities.  Ignored unless \code{data} is a data frame.}

\item{betaRef}{Indices of reference elements for t.p.m. multinomial logit link. Ignored unless \code{data} is a data frame.}

\item{stateNames}{Optional character vector of length nbStates indicating state names. Ignored unless \code{data} is a data frame.}

\item{hierStates}{A hierarchical model structure \code{\link[data.tree]{Node}} for the states ('state').  See \code{\link{fitHMM}}.}

\item{hierBeta}{A hierarchical data structure \code{\link[data.tree]{Node}} for the matrix of regression coefficients for the transition probabilities at each level of the hierarchy, including initial values ('beta'), parameter equality constraints ('betaCons'), fixed parameters ('fixPar'), and working scale bounds ('workBounds'). See details.}

\item{hierFormula}{A hierarchical formula structure for the transition probability covariates for each level of the hierarchy ('formula'). See \code{\link{fitHMM}}.}

\item{hierDist}{A hierarchical data structure \code{\link[data.tree]{Node}} for the data streams ('dist'). See \code{\link{fitHMM}}.}
}
\value{
If \code{mixtures=1}, an array of dimension \code{nbStates} x \code{nbStates} x \code{nrow(data)} containing the t.p.m for each observation in \code{data}.
If \code{mixtures>1}, a list of length \code{mixtures}, where each element is an array of dimension \code{nbStates} x \code{nbStates} x \code{nrow(data)} containing the t.p.m for each observation in \code{data}.

If a hierarchical HMM structure is provided, then a hierarchical data structure containing the state transition probabilities for each time step at each level of the hierarchy ('gamma') is returned.
}
\description{
Computation of the transition probability matrix for each time step as a function of the covariates and the regression
parameters.
}
\examples{
m <- example$m
trProbs <- getTrProbs(m)

# equivalent
trProbs <- getTrProbs(m$data,nbStates=2,beta=m$mle$beta,formula=m$conditions$formula)

}
