% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMfits.R
\name{HMMfits}
\alias{HMMfits}
\title{Constructor of \code{HMMfits} objects}
\usage{
HMMfits(m)
}
\arguments{
\item{m}{A list of \code{\link{momentuHMM}} objects.

\code{HMMfits} objects are returned by \code{\link{MIfitHMM}} when arguments \code{fit=TRUE} and \code{poolEstimates=FALSE}.}
}
\value{
An object \code{HMMfits}.
}
\description{
Constructor of \code{HMMfits} objects
}
