% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crwData.R
\name{is.crwData}
\alias{is.crwData}
\title{Is crwData}
\usage{
is.crwData(x)
}
\arguments{
\item{x}{An R object}
}
\value{
\code{TRUE} if \code{x} is of class \code{\link{crwData}}, \code{FALSE} otherwise.
}
\description{
Check that an object is of class \code{\link{crwData}}. Used in \code{\link{MIfitHMM}}.
}
