% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crwHierData.R
\name{is.crwHierData}
\alias{is.crwHierData}
\title{Is crwHierData}
\usage{
is.crwHierData(x)
}
\arguments{
\item{x}{An R object}
}
\value{
\code{TRUE} if \code{x} is of class \code{\link{crwHierData}}, \code{FALSE} otherwise.
}
\description{
Check that an object is of class \code{\link{crwHierData}}. Used in \code{\link{MIfitHMM}}.
}
