% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crwHierSim.R
\name{is.crwHierSim}
\alias{is.crwHierSim}
\title{Is crwHierSim}
\usage{
is.crwHierSim(x)
}
\arguments{
\item{x}{An R object}
}
\value{
\code{TRUE} if \code{x} is of class \code{\link{crwHierSim}}, \code{FALSE} otherwise.
}
\description{
Check that an object is of class \code{\link{crwHierSim}}.
}
