% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crwSim.R
\name{is.crwSim}
\alias{is.crwSim}
\title{Is crwSim}
\usage{
is.crwSim(x)
}
\arguments{
\item{x}{An R object}
}
\value{
\code{TRUE} if \code{x} is of class \code{\link{crwSim}}, \code{FALSE} otherwise.
}
\description{
Check that an object is of class \code{\link{crwSim}}.
}
