% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMfits.R
\name{is.HMMfits}
\alias{is.HMMfits}
\title{Is HMMfits}
\usage{
is.HMMfits(x)
}
\arguments{
\item{x}{An R object}
}
\value{
\code{TRUE} if \code{x} is of class \code{\link{HMMfits}}, \code{FALSE} otherwise.
}
\description{
Check that an object is of class \code{\link{HMMfits}}.
}
