% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miHMM.R
\name{is.miHMM}
\alias{is.miHMM}
\title{Is miHMM}
\usage{
is.miHMM(x)
}
\arguments{
\item{x}{An R object}
}
\value{
\code{TRUE} if \code{x} is of class \code{\link{miHMM}}, \code{FALSE} otherwise.
}
\description{
Check that an object is of class \code{\link{miHMM}}.
}
