% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miSum.R
\name{is.miSum}
\alias{is.miSum}
\title{Is miSum}
\usage{
is.miSum(x)
}
\arguments{
\item{x}{An R object}
}
\value{
\code{TRUE} if \code{x} is of class \code{\link{miSum}}, \code{FALSE} otherwise.
}
\description{
Check that an object is of class \code{\link{miSum}}.
}
