% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentuHierHMM.R
\name{is.momentuHierHMM}
\alias{is.momentuHierHMM}
\title{Is momentuHierHMM}
\usage{
is.momentuHierHMM(x)
}
\arguments{
\item{x}{An R object}
}
\value{
\code{TRUE} if \code{x} is of class \code{\link{momentuHierHMM}}, \code{FALSE} otherwise.
}
\description{
Check that an object is of class \code{\link{momentuHierHMM}}. Used in \code{\link{CIreal}}, \code{\link{CIbeta}},
\code{\link{plotPR}}, \code{\link{plotStates}}, \code{\link{pseudoRes}}, \code{\link{stateProbs}},
and \code{\link{viterbi}}.
}
