% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentuHierHMMData.R
\name{is.momentuHierHMMData}
\alias{is.momentuHierHMMData}
\title{Is momentuHierHMMData}
\usage{
is.momentuHierHMMData(x)
}
\arguments{
\item{x}{An R object}
}
\value{
\code{TRUE} if \code{x} is of class \code{\link{momentuHierHMMData}}, \code{FALSE} otherwise.
}
\description{
Check that an object is of class \code{\link{momentuHierHMMData}}. Used in \code{\link{fitHMM}}.
}
