% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentuHMM.R
\name{is.momentuHMM}
\alias{is.momentuHMM}
\title{Is momentuHMM}
\usage{
is.momentuHMM(x)
}
\arguments{
\item{x}{An R object}
}
\value{
\code{TRUE} if \code{x} is of class \code{\link{momentuHMM}}, \code{FALSE} otherwise.
}
\description{
Check that an object is of class \code{\link{momentuHMM}}. Used in \code{\link{CIreal}}, \code{\link{CIbeta}},
\code{\link{plotPR}}, \code{\link{plotStates}}, \code{\link{pseudoRes}}, \code{\link{stateProbs}},
and \code{\link{viterbi}}.
}
