% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentuHMMData.R
\name{is.momentuHMMData}
\alias{is.momentuHMMData}
\title{Is momentuHMMData}
\usage{
is.momentuHMMData(x)
}
\arguments{
\item{x}{An R object}
}
\value{
\code{TRUE} if \code{x} is of class \code{\link{momentuHMMData}}, \code{FALSE} otherwise.
}
\description{
Check that an object is of class \code{\link{momentuHMMData}}. Used in \code{\link{fitHMM}}.
}
