% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentuHMMMI.R
\name{is.momentuHMMMI}
\alias{is.momentuHMMMI}
\title{Is momentuHMMMI}
\usage{
is.momentuHMMMI(x)
}
\arguments{
\item{x}{An R object}
}
\value{
\code{TRUE} if \code{x} is of class \code{\link{momentuHMMMI}}, \code{FALSE} otherwise.
}
\description{
Check that an object is of class \code{\link{momentuHMMMI}}.
}

