% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveData.R
\name{is.moveData}
\alias{is.moveData}
\title{Is moveData}
\usage{
is.moveData(x)
}
\arguments{
\item{x}{An R object}
}
\value{
\code{TRUE} if \code{x} is of class \code{\link{moveData}}, \code{FALSE} otherwise.
}
\description{
Check that an object is of class \code{\link{moveData}}. Used in \code{\link{fitHMM}}.
}

