% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miSum.R
\name{miSum}
\alias{miSum}
\title{Constructor of \code{miSum} objects}
\usage{
miSum(m)
}
\arguments{
\item{m}{A list of attributes of the multiple imputation model: \code{Par} (the multiple imputation estimates of
the parameters of the model), \code{data} (the movement data), \code{conditions} (conditions used to fit
the model)}
}
\value{
An object \code{momentuHMM}.
}
\description{
Constructor of \code{miSum} objects
}

