% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentuHierHMM.R
\name{momentuHierHMM}
\alias{momentuHierHMM}
\title{Constructor of \code{momentuHierHMM} objects}
\usage{
momentuHierHMM(m)
}
\arguments{
\item{m}{A list of attributes of the fitted model: \code{mle} (the maximum likelihood estimates of
the parameters of the model), \code{data} (the \code{fitHMM} data), \code{mod} (the object
returned by the \code{fitHMM} numerical optimizer \code{nlm} or \code{optim}), \code{conditions} (conditions used to fit
the model: \code{hierStates}, \code{hierDist}, \code{zeroInflation}, \code{estAngleMean}, \code{circularAngleMean}
\code{stationary}, \code{formula}, \code{userBounds}, \code{bounds}, \code{workBounds}, \code{DM}, etc.), 
\code{stateNames}, and \code{rawCovs} (optional -- only if there are transition probability matrix covariates in the data).}
}
\value{
An object \code{momentuHierHMM}.
}
\description{
Constructor of \code{momentuHierHMM} objects
}
