% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentuHierHMMData.R
\name{momentuHierHMMData}
\alias{momentuHierHMMData}
\title{Constructor of \code{momentuHierHMMData} objects}
\usage{
momentuHierHMMData(data)
}
\arguments{
\item{data}{A dataframe containing: \code{ID} (the ID(s) of the observed animal(s)), \code{level} (the level of the hierarchy for each observation), and the data streams such as \code{step}
(the step lengths, if any), \code{angle} (the turning angles, if any), \code{x} (either easting or longitude, if any),
\code{y} (either norting or latitude, if any), and covariates (if any).}
}
\value{
An object \code{momentuHierHMMData}.
}
\description{
Constructor of \code{momentuHierHMMData} objects
}
