% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentuHMM.R
\name{momentuHMM}
\alias{momentuHMM}
\title{Constructor of \code{momentuHMM} objects}
\usage{
momentuHMM(m)
}
\arguments{
\item{m}{A list of attributes of the fitted model: \code{mle} (the maximum likelihood estimates of
the parameters of the model), \code{data} (the movement data), \code{mod} (the object
returned by the numerical optimizer \code{nlm}), \code{conditions} (a few conditions used to fit
the model: \code{stepDist}, \code{angleDist}, \code{zeroInflation}, \code{estAngleMean},
\code{stationary}, and \code{formula}), \code{rawCovs} (optional -- only if there are covariates
in the data).}
}
\value{
An object \code{momentuHMM}.
}
\description{
Constructor of \code{momentuHMM} objects
}
