% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentuHMMMI.R
\name{momentuHMMMI}
\alias{momentuHMMMI}
\title{Constructor of \code{momentuHMMMI} objects}
\usage{
momentuHMMMI(m)
}
\arguments{
\item{m}{A list of attributes of the multiple imputation model: \code{Par} (the multiple imputation estimates of
the parameters of the model), \code{data} (the movement data), \code{conditions} (conditions used to fit
the model), \code{bounds} (parameter bounds), \code{rawCovs} (optional -- only if there are covariates
in the data).}
}
\value{
An object \code{momentuHMM}.
}
\description{
Constructor of \code{momentuHMMMI} objects
}

