% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveData.R
\name{moveData}
\alias{moveData}
\title{Constructor of \code{moveData} objects}
\usage{
moveData(data)
}
\arguments{
\item{data}{A dataframe containing: \code{ID} (the ID(s) of the observed animal(s)), \code{step}
(the step lengths), \code{angle} (the turning angles, if any), \code{x} (either easting or longitude),
\code{y} (either norting or latitude), and covariates, if any.}
}
\value{
An object \code{moveData}.
}
\description{
Constructor of \code{moveData} objects
}

