% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nLogLike_rcpp}
\alias{nLogLike_rcpp}
\title{Negative log-likelihood}
\usage{
nLogLike_rcpp(nbStates, covs, data, dataNames, dist, Par, aInd, zeroInflation,
  stationary, knownStates)
}
\arguments{
\item{nbStates}{Number of states}

\item{covs}{Covariates}

\item{data}{A \code{\link{momentuHMMData}} object of the observations}

\item{aInd}{Vector of indices of the rows at which the data switches to another animal}

\item{zeroInflation}{\code{true} if zero-inflation is included in the step length distribution,
\code{false} otherwise.}

\item{stationary}{\code{false} if there are covariates. If \code{true}, the initial distribution is considered
equal to the stationary distribution. Default: \code{false}.}

\item{knownStates}{Vector of values of the state process which are known prior to fitting the
model (if any). Default: NULL (states are not known). This should be a vector with length the number
of rows of 'data'; each element should either be an integer (the value of the known states) or NA if
the state is not known.}

\item{beta}{Matrix of regression coefficients for the transition probabilities}

\item{stepDist}{The name of the step length distribution}

\item{angleDist}{The name of the turning angle distribution}

\item{stepPar}{State-dependent parameters of the step length distribution}

\item{anglePar}{State-dependent parameters of the turning angle distribution}

\item{delta}{Stationary distribution}
}
\value{
Negative log-likelihood
}
\description{
Computation of the negative log-likelihood (forward algorithm - written in C++)
}
