% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_momentuHMMData.R
\name{plot.momentuHMMData}
\alias{plot.momentuHMMData}
\alias{plot.momentuHierHMMData}
\title{Plot \code{momentuHMMData} or \code{momentuHierHMMData}}
\usage{
\method{plot}{momentuHMMData}(
  x,
  dataNames = c("step", "angle"),
  animals = NULL,
  compact = FALSE,
  ask = TRUE,
  breaks = "Sturges",
  ...
)
}
\arguments{
\item{x}{An object \code{momentuHMMData} or \code{momentuHierHMMData}}

\item{dataNames}{Names of the variables to plot. Default is \code{dataNames=c("step","angle")}.}

\item{animals}{Vector of indices or IDs of animals for which information will be plotted.
Default: \code{NULL} ; all animals are plotted.}

\item{compact}{\code{TRUE} for a compact plot (all individuals at once), \code{FALSE} otherwise
(default -- one individual at a time).}

\item{ask}{If \code{TRUE}, the execution pauses between each plot.}

\item{breaks}{Histogram parameter. See \code{hist} documentation.}

\item{...}{Currently unused. For compatibility with generic method.}
}
\description{
Plot \code{momentuHMMData} or \code{momentuHierHMMData}
}
\examples{
# data is a momentuHMMData object (as returned by prepData), automatically loaded with the package
data <- example$m$data

plot(data,dataNames=c("step","angle","cov1","cov2"),
     compact=TRUE,breaks=20,ask=FALSE)

}
