% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\title{Preprocessing of the tracking data}
\usage{
prepData(Data, type = c("LL", "UTM"), coordNames = NULL, covNames = NULL)
}
\arguments{
\item{Data}{A dataframe of data streams, including optionally a field \code{ID}
(identifiers for the observed individuals), coordinates from which step length ('step') and turning angle ('angle') area calculated, and \code{covNames} that identify covariates.  If coordinates are included. If step length and turning are to be calculated from coordinates, the \code{coordNames} argument must identify the names for the longitunal and latitudinal coordinates (e.g., \code{coordNames=c("x","y")}).}

\item{type}{\code{'LL'} if longitude/latitude provided (default), \code{'UTM'} if easting/northing.}

\item{coordNames}{Names of the columns of coordinates in the data frame. Default: \code{c("x","y")}.}
}
\value{
An object \code{moveData}, i.e. a dataframe of:
\item{ID}{The ID(s) of the observed animal(s)}
\item{step}{The step lengths - in kilometers if longitude/latitude provided, and in the metrics of
the data otherwise}
\item{angle}{The turning angles (if any) - in radians}
\item{x}{Either easting or longitude}
\item{y}{Either norting or latitude}
\item{...}{Covariates (if any)}
}
\description{
Preprocessing of the tracking data
}
\examples{
coord1 <- c(1,2,3,4,5,6,7,8,9,10)
coord2 <- c(1,1,1,2,2,2,1,1,1,2)
Data <- data.frame(coord1=coord1,coord2=coord2)
d <- prepData(Data,type='UTM',coordNames=c("coord1","coord2"))

}

