% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_miHMM.R
\name{print.miHMM}
\alias{print.miHMM}
\title{Print \code{miHMM}}
\usage{
\method{print}{miHMM}(x, ...)
}
\arguments{
\item{x}{A \code{miHMM} object.}

\item{...}{Currently unused. For compatibility with generic method.}
}
\description{
Print \code{miHMM}
}
\examples{
# m is a miHMM object (as returned by MIfitHMM), automatically loaded with the package
m <- example$m

print(m)

}

