% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_miSum.R
\name{print.miSum}
\alias{print.miSum}
\title{Print \code{miSum}}
\usage{
\method{print}{miSum}(x, ...)
}
\arguments{
\item{x}{A \code{miSum} object.}

\item{...}{Currently unused. For compatibility with generic method.}
}
\description{
Print \code{miSum}
}
\examples{
# create miSum object from example data
mi <- MIpool(miExample$HMMfits,ncores=1)
print(mi)

}
