% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_momentuHMM.R
\name{print.momentuHMM}
\alias{print.momentuHMM}
\alias{print.momentuHierHMM}
\title{Print \code{momentuHMM}}
\usage{
\method{print}{momentuHMM}(x, ...)

\method{print}{momentuHierHMM}(x, ...)
}
\arguments{
\item{x}{A \code{momentuHMM} object.}

\item{...}{Currently unused. For compatibility with generic method.}
}
\description{
Print \code{momentuHMM}
}
\examples{
# m is a momentuHMM object (as returned by fitHMM), automatically loaded with the package
m <- example$m

print(m)

}
