% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_momentuHMMMI.R
\name{print.momentuHMMMI}
\alias{print.momentuHMMMI}
\title{Print \code{momentuHMMMI}}
\usage{
\method{print}{momentuHMMMI}(x, ...)
}
\arguments{
\item{x}{A \code{momentuHMMMI} object.}

\item{...}{Currently unused. For compatibility with generic method.}
}
\description{
Print \code{momentuHMMMI}
}

