% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setModelName.R
\name{setModelName}
\alias{setModelName}
\title{Set \code{modelName} for a \code{momentuHMM}, \code{miHMM}, \code{HMMfits}, or \code{miSum} object}
\usage{
setModelName(model, modelName)
}
\arguments{
\item{model}{\code{\link{momentuHMM}}, \code{\link{miHMM}}, \code{\link{HMMfits}}, or \code{\link{miSum}} object}

\item{modelName}{Character string providing a name for the model. See \code{\link{fitHMM}} and \code{\link{MIfitHMM}}.}
}
\value{
\code{model} object with new \code{modelName} field
}
\description{
Set \code{modelName} for a \code{momentuHMM}, \code{miHMM}, \code{HMMfits}, or \code{miSum} object
}
\examples{
m <- example$m
mName <- setModelName(m, modelName="example")

}
