% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setStateNames.R
\name{setStateNames}
\alias{setStateNames}
\title{Set \code{stateNames} for a \code{momentuHMM}, \code{miHMM}, \code{HMMfits}, or \code{miSum} object}
\usage{
setStateNames(model, stateNames)
}
\arguments{
\item{model}{\code{\link{momentuHMM}}, \code{\link{miHMM}}, \code{\link{HMMfits}}, or \code{\link{miSum}} object}

\item{stateNames}{Character string providing state names for the model. See \code{\link{fitHMM}} and \code{\link{MIfitHMM}}.}
}
\value{
\code{model} object with new \code{stateNames} field
}
\description{
Set \code{stateNames} for a \code{momentuHMM}, \code{miHMM}, \code{HMMfits}, or \code{miSum} object
}
\examples{
m <- example$m
mName <- setStateNames(m, stateNames=c("encamped","exploratory"))

}
