% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stationary_rcpp}
\alias{stationary_rcpp}
\title{Stationary distribution for a continuous-time Markov chain}
\usage{
stationary_rcpp(A)
}
\arguments{
\item{A}{transition rate matrix (of dimension nbStates x nbStates)}
}
\value{
row vector of stationary distribution probabilities
}
\description{
Written in C++.
}
