% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_momentuHMMData.R
\name{summary.momentuHMMData}
\alias{summary.momentuHMMData}
\alias{summary.momentuHierHMMData}
\title{Summary \code{momentuHMMData}}
\usage{
\method{summary}{momentuHMMData}(object, dataNames = c("step", "angle"), animals = NULL, ...)

\method{summary}{momentuHierHMMData}(object, dataNames = c("step", "angle", "level"), animals = NULL, ...)
}
\arguments{
\item{object}{A \code{\link{momentuHMMData}} or \code{\link{momentuHierHMMData}} object.}

\item{dataNames}{Names of the variables to summarize. Default is \code{dataNames=c("step","angle")}.}

\item{animals}{Vector of indices or IDs of animals for which data will be summarized.
Default: \code{NULL} ; data for all animals are summarized.}

\item{...}{Currently unused. For compatibility with generic method.}
}
\description{
Summary \code{momentuHMMData}
}
\examples{
# data is a momentuHMMData object (as returned by prepData), automatically loaded with the package
data <- example$m$data

summary(data,dataNames=c("step","angle","cov1","cov2"))

}
