% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_moveData.R
\name{summary.moveData}
\alias{summary.moveData}
\title{Summary \code{moveData}}
\usage{
\method{summary}{moveData}(object, details = TRUE, ...)
}
\arguments{
\item{object}{A \code{moveData} object.}

\item{details}{\code{TRUE} if quantiles of the covariate values should be printed (default),
\code{FALSE} otherwise.}

\item{...}{Currently unused. For compatibility with generic method.}
}
\description{
Summary \code{moveData}
}
\examples{
# m is a moveData object (as returned by prepData), automatically loaded with the package
data <- example$data

summary(data)

}

