% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trMatrix_rcpp}
\alias{trMatrix_rcpp}
\title{Transition probability matrix}
\usage{
trMatrix_rcpp(nbStates, beta, covs, betaRef, CT = FALSE, dt = as.numeric(c()))
}
\arguments{
\item{nbStates}{Number of states}

\item{beta}{Matrix of regression parameters}

\item{covs}{Matrix of covariate values}

\item{betaRef}{Indices of reference elements for t.p.m. multinomial logit link.}

\item{CT}{logical indicating discrete-time approximation of a continuous-time model}

\item{dt}{numeric vector of length \code{nrow(covs)} indicating the time difference between observations. Ignored unless \code{CT=TRUE}.}
}
\value{
Three dimensional array \code{trMat}, such that \code{trMat[,,t]} is the transition matrix at
time t.
}
\description{
Computation of the transition probability matrix, as a function of the covariates and the regression
parameters. Written in C++. Used in \code{\link{viterbi}}.
}
