% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{XBloop_rcpp}
\alias{XBloop_rcpp}
\title{Get XB}
\usage{
XBloop_rcpp(DM, Xvec, nbObs, nr, nc)
}
\arguments{
\item{DM}{design matrix}

\item{Xvec}{working parameters}

\item{nbObs}{number of observations}

\item{nr}{number of rows in design matrix}

\item{nc}{number of column in design matrix}
}
\value{
XB matrix
}
\description{
Loop for computation of design matrix (X) times the working scale parameters (B). Written in C++. Used in \code{\link{w2n}}.
}
