% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allProbs.R
\name{allProbs}
\alias{allProbs}
\title{Matrix of all probabilities}
\usage{
allProbs(m, nbStates)
}
\arguments{
\item{m}{Object \code{momentuHMM}.}

\item{nbStates}{Number of states of the HMM.}
}
\value{
Matrix of all probabilities.
}
\description{
Used in functions \code{\link{viterbi}}, \code{\link{logAlpha}}, \code{\link{logBeta}}.
}
\examples{
\dontrun{
P <- momentuHMM:::allProbs(m=example$m,nbStates=2)
}
}
