% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmvnorm_rcpp}
\alias{dmvnorm_rcpp}
\title{C++ implementation of multivariate Normal probability density function for multiple inputs}
\usage{
dmvnorm_rcpp(x, mean, varcovM)
}
\arguments{
\item{x}{data matrix of dimension \code{p x n}, \code{p} being the dimension of the
data and n the number of data points.}

\item{mean}{mean vectors matrix of dimension \code{p x n}}

\item{varcovM}{list of length \code{n} of variance-covariance matrices,
each of dimensions \code{p x p}.}
}
\value{
matrix of densities of dimension \code{K x n}.
}
\description{
C++ implementation of multivariate Normal probability density function for multiple inputs
}
