% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exGen.R
\docType{data}
\name{example}
\alias{example}
\alias{miExample}
\alias{forest}
\title{Example dataset}
\usage{
example
}
\description{
These data are generated by the function \code{\link{exGen}}, and used in the examples and tests of
other functions to keep them as short as possible.
}
\details{
\code{example} is a list of the following objects for demonstrating \code{\link{fitHMM}}:
\itemize{
\item \code{m} A \code{\link{momentuHMM}} object
\item \code{simPar} The parameters used to simulate \code{data}
\item \code{par0} The initial parameters in the optimization to fit \code{m}
}

\code{miExample} is a list of the following objects for demonstrating \code{\link{crawlWrap}}, \code{\link{MIfitHMM}}, and \code{\link{MIpool}}:
\itemize{
\item \code{obsData} Simulated observation data with measurement error and temporal irregularity (generated by \code{\link{simData}})
\item \code{inits} initial states for \code{crawlWrap} example
\item \code{err.model} Error ellipse model for \code{crawlWrap} example
}

\code{forest} is a simulated spatial covariate raster layer
}
