% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expmatrix_rcpp}
\alias{expmatrix_rcpp}
\title{Matrix Exponential}
\usage{
expmatrix_rcpp(x)
}
\arguments{
\item{x}{a square matrix.}
}
\value{
The matrix exponential of x.
}
\description{
This function computes the exponential of a square matrix \code{A}, defined as the sum from \code{r=0} to infinity of \code{A^r/r!}, using the default method of \code{\link[expm]{expm}}.
}
