% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expmatrix_rcpp}
\alias{expmatrix_rcpp}
\title{Matrix Exponential}
\usage{
expmatrix_rcpp(x, maxRate = Inf, check = FALSE)
}
\arguments{
\item{x}{a square matrix.}

\item{maxRate}{maximum allowable value for the off-diagonal state transition rate parameters. Default: \code{Inf}. Setting less than \code{Inf} can help avoid numerical issues during optimization.}

\item{check}{logical indicating whether or not to check transition probability matrix for issues}
}
\value{
The matrix exponential of x.
}
\description{
This function computes the exponential of a square matrix \code{A}, defined as the sum from \code{r=0} to infinity of \code{A^r/r!}, using the default method of \code{\link[expm]{expm}}.
}
