% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrProbs.R
\name{getTrProbs}
\alias{getTrProbs}
\alias{getTrProbs.default}
\alias{getTrProbs.hierarchical}
\title{Transition probability matrix}
\usage{
getTrProbs(data, ...)

\method{getTrProbs}{default}(data, nbStates, beta, workBoundsBeta = NULL,
  formula = ~1, mixtures = 1, betaRef = NULL, stateNames = NULL, ...)

\method{getTrProbs}{hierarchical}(data, hierStates, beta,
  workBoundsBeta = NULL, hierFormula = NULL, mixtures = 1, hierDist, ...)
}
\arguments{
\item{data}{\code{\link{momentuHMMData}} object, \code{\link{momentuHierHMMData}} object, or a data frame containing the covariate values. 

If a data frame is provided, then either \code{nbStates} must be specified (for a regular HMM) or \code{hierStates} and \code{hierDist}
must be specified (for a hierarchical HMM).}

\item{...}{further arguments passed to or from other methods}

\item{nbStates}{Number of states}

\item{beta}{Matrix of regression coefficients for the transition probabilities}

\item{workBoundsBeta}{An optional 2-column matrix specifying bounds on the working scale of the transition probability parameters. Must be a k x 2 matrix, where k=\code{length(beta)}.
The first column pertains to the lower bound and the second column the upper bound.}

\item{formula}{Regression formula for the transition probability covariates.}

\item{mixtures}{Number of mixtures for the state transition probabilities}

\item{betaRef}{Indices of reference elements for t.p.m. multinomial logit link.}

\item{stateNames}{Optional character vector of length nbStates indicating state names.}

\item{hierStates}{A hierarchical model structure \code{\link[data.tree]{Node}} for the states ('state').  See \code{\link{fitHMM}}.}

\item{hierFormula}{A hierarchical formula structure for the transition probability covariates for each level of the hierarchy ('formula'). See \code{\link{fitHMM}}.}

\item{hierDist}{A hierarchical data structure \code{\link[data.tree]{Node}} for the data streams ('dist'). See \code{\link{fitHMM}}.}
}
\value{
If \code{mixtures=1}, an array of dimension \code{nbStates} x \code{nbStates} x \code{nrow(data)} containing the t.p.m for each observation in \code{data}.
If \code{mixtures>1}, a list of length \code{mixtures}, where each element is an array of dimension \code{nbStates} x \code{nbStates} x \code{nrow(data)} containing the t.p.m for each observation in \code{data}.
}
\description{
Computation of the transition probability matrix for each time step as a function of the covariates and the regression
parameters.
}
\examples{
m <- example$m
trProbs <- getTrProbs(m$data,nbStates=2,beta=m$mle$beta,formula=m$conditions$formula)

}
