% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miHMM.R
\name{miHMM}
\alias{miHMM}
\title{Constructor of \code{miHMM} objects}
\usage{
miHMM(m)
}
\arguments{
\item{m}{A list with attributes \code{miSum} (a \code{\link{miSum}} object) and \code{HMMfits} (a list of \code{\link{momentuHMM}} objects).

\code{miHMM} objects are returned by \code{\link{MIfitHMM}} when arguments \code{fit=TRUE}, \code{nSims>1}, and \code{poolEstimates=TRUE}.}
}
\value{
An object \code{miHMM}.
}
\description{
Constructor of \code{miHMM} objects
}
