% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miSum.R
\name{miSum}
\alias{miSum}
\title{Constructor of \code{miSum} objects}
\usage{
miSum(m)
}
\arguments{
\item{m}{A list of attributes required for multiple imputation summaries: \code{data} (averaged across imputations), \code{Par} (the pooled estimates of
the parameters of the model), \code{conditions} (conditions used to fit the model), and \code{MIcombine} (as returned by \code{\link[mitools]{MIcombine}} for
the working parameters).}
}
\value{
An object \code{miSum}.
}
\description{
Constructor of \code{miSum} objects
}
