% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentuHMM.R
\name{momentuHMM}
\alias{momentuHMM}
\title{Constructor of \code{momentuHMM} objects}
\usage{
momentuHMM(m)
}
\arguments{
\item{m}{A list of attributes of the fitted model: \code{mle} (the maximum likelihood estimates of
the parameters of the model), \code{data} (the \code{fitHMM} data), \code{mod} (the object
returned by the \code{fitHMM} numerical optimizer \code{nlm}), \code{conditions} (conditions used to fit
the model: \code{dist}, \code{zeroInflation}, \code{estAngleMean}, \code{circularAngleMean}
\code{stationary}, \code{formula}, \code{cons}, \code{userBounds}, \code{bounds}, \code{workcons}, \code{DM}, etc.), 
\code{stateNames}, and \code{rawCovs} (optional -- only if there are transition probability matrix covariates in the data).}
}
\value{
An object \code{momentuHMM}.
}
\description{
Constructor of \code{momentuHMM} objects
}
