% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nLogLike_rcpp}
\alias{nLogLike_rcpp}
\title{Negative log-likelihood}
\usage{
nLogLike_rcpp(nbStates, beta, covs, data, stepDist, angleDist, omegaDist,
  dryDist, diveDist, iceDist, landDist, stepPar, anglePar, omegaPar, dryPar,
  divePar, icePar, landPar, delta, aInd, zeroInflation = FALSE,
  stationary = FALSE)
}
\arguments{
\item{nbStates}{Number of states}

\item{beta}{Matrix of regression coefficients for the transition probabilities}

\item{covs}{Covariates}

\item{data}{A \code{\link{moveData}} object of the observations}

\item{stepDist}{The name of the step length distribution}

\item{angleDist}{The name of the turning angle distribution}

\item{stepPar}{State-dependent parameters of the step length distribution}

\item{anglePar}{State-dependent parameters of the turning angle distribution}

\item{delta}{Stationary distribution}

\item{aInd}{Vector of indices of the rows at which the data switches to another animal}

\item{zeroInflation}{\code{true} if zero-inflation is included in the step length distribution,
\code{false} otherwise.}

\item{stationary}{\code{false} if there are covariates. If \code{true}, the initial distribution is considered
equal to the stationary distribution. Default: \code{false}.}
}
\value{
Negative log-likelihood
}
\description{
Computation of the negative log-likelihood (forward algorithm - written in C++)
}

