% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepCTDS.R
\name{prepCTDS}
\alias{prepCTDS}
\alias{prepCTDS.default}
\alias{prepCTDS.hierarchical}
\title{Preprocessing of continuous-time discrete-space (CTDS) movement HMMs using ctmcmove}
\usage{
prepCTDS(data, ...)

\method{prepCTDS}{default}(
  data,
  Time.unit = "auto",
  rast,
  directions = 4,
  zero.idx = integer(),
  print.iter = FALSE,
  interpMethod = "ShortestPath",
  spatialCovs = NULL,
  spatialCovs.grad = NULL,
  crw = TRUE,
  normalize.gradients = FALSE,
  grad.point.decreasing = FALSE,
  covNames = NULL,
  ncores = 1,
  ...
)

\method{prepCTDS}{hierarchical}(
  data,
  Time.unit = "auto",
  rast,
  directions = 4,
  zero.idx = integer(),
  print.iter = FALSE,
  interpMethod = "ShortestPath",
  spatialCovs = NULL,
  spatialCovs.grad = NULL,
  crw = TRUE,
  normalize.gradients = FALSE,
  grad.point.decreasing = FALSE,
  covNames = NULL,
  ncores = 1,
  hierLevels,
  coordLevel,
  ...
)
}
\arguments{
\item{data}{Either a \code{data.frame} of data streams or a \code{\link{crwData}} (or \code{\link{crwHierData}}) object (as returned by \code{\link{crawlWrap}}). If a \code{data.frame}, it must include entries for the x coordinate (\code{x}), y coordinate (\code{y}), and time stamp (\code{time}). An \code{ID} entry must also be included if \code{data} includes multiple individuals.}

\item{...}{further arguments passed to or from other methods}

\item{Time.unit}{Character string indicating units for time difference between observations (e.g. 'auto', 'secs', 'mins', 'hours', 'days', 'weeks'). Ignored unless \code{data$time} is of class \code{\link[base]{date-time}} or \code{\link[base]{date}}. Default: 'auto', but note that if there are multiple individuals, then the units are determined based on the time stamps for the first individual.}

\item{rast}{A raster object or raster stack object that will define the discrete-space grid cells for the CTMC movement path.}

\item{directions}{Integer. Either 4 (indicating a "Rook's neighborhood" of 4 neighboring grid cells) or 8 (indicating a "King's neighborhood" of 8 neighboring grid cells).}

\item{zero.idx}{Integer vector of the indices of raster cells that are not passable and should be excluded. These are cells where movement should be impossible. Default is zero.idx=integer().}

\item{print.iter}{Logical. If true, then the progress stepping through each observed location in \code{data} will be output in the terminal.}

\item{interpMethod}{Specifies interpolation method. Either "ShortestPath", which uses the shortest graphical path on the raster graph, or "LinearInterp", which linearly interpolates between observed locations. "ShortestPath" is slower, slightly more accurate, and allows for impassible barriers specified through "zero.idx". "LinearInterp" is faster but does not allow for impassible barriers.}

\item{spatialCovs}{List of \code{\link[raster]{raster}} objects for spatio-temporally referenced covariates. Covariates specified by \code{spatialCovs} are extracted from the raster 
layer(s) based on the location data (and the z values for a raster \code{\link[raster]{stack}} 
or \code{\link[raster]{brick}}) for each time step.  If an element of \code{spatialCovs} is a raster \code{\link[raster]{stack}} or \code{\link[raster]{brick}}, 
then z values must be set using \code{raster::setZ} and \code{data} must include column(s) of the corresponding z value(s) for each observation (e.g., 'Date').}

\item{spatialCovs.grad}{List of \code{\link[raster]{raster}} objects for spatio-temporally referenced covariates, where a directional gradient is to be calculated internally using \code{\link[ctmcmove]{rast.grad}}. Gradient-based covariates specified by \code{spatialCovs.grad} are extracted from the raster 
layer(s) based on the location data (and the z values for a raster \code{\link[raster]{stack}} or \code{\link[raster]{brick}}) for each time step.  If an element of \code{spatialCovs.grad} is a raster \code{\link[raster]{stack}} or \code{\link[raster]{brick}}, 
then z values must be set using \code{raster::setZ} and \code{data} must include column(s) of the corresponding z value(s) for each observation (e.g., 'Date').}

\item{crw}{Logical. If TRUE (default), an autocovariate is created for each cell visited in the CTMC movement path. The autocovariate is a unit-length directional vector pointing from the center of the most recent grid cell to the center of the current grid cell.}

\item{normalize.gradients}{Logical. Default is FALSE. If TRUE, then all gradient covariates for \code{spatialCovs.grad} are normalized by dividing by the length of the gradient vector at each point.}

\item{grad.point.decreasing}{Logical. Default is FALSE. If TRUE, then the gradient covariates are positive in the direction of decreasing values of the covariate. If FALSE, then the gradient covariates are positive in the direction of increasing values of the covariate (like a true gradient).}

\item{covNames}{Character vector indicating the names of any covariates in \code{data}. Any variables in \code{data} (other than \code{ID}, \code{x}, \code{y}, and \code{time}) that are not identified in covNames are assumed to be additional data streams (i.e., missing values will not be accounted for).}

\item{ncores}{Number of cores to use for parallel processing. Default: 1 (no parallel processing).}

\item{hierLevels}{Character vector indicating the levels of the hierarchy and their order, from top (coarsest scale) to bottom (finest scale), that are included in \code{data$level}. For example, for a 2-level hierarchy then 
\code{hierLevels=c("1","2i","2")} indicates \code{data$level} for each observation can be one of three factor levels: "1" (coarse scale), "2i" (initial fine scale), and "2" (fine scale).  Ignored if \code{data} is a \code{\link{crwHierData}} object.}

\item{coordLevel}{Character string indicating the level of the hierarchy for the location data. If specified, then \code{data} must include a 'level' field indicating the level of the hierarchy for each observation.  Ignored if \code{coordNames} is \code{NULL} or \code{data} is a \code{\link{crwHierData}} object.}
}
\value{
A \code{\link{momentuHMMData}} object of class \code{ctds} that can be passed to \code{\link{fitCTHMM}} (or as a list to \code{\link{MIfitCTHMM}}), where:
\item{ID}{The ID(s) of the observed animal(s)}
\item{time}{Time stamp for each observation}
\item{x}{Easting coordinate of the current cell}
\item{y}{Northing coordinate of the current cell}
\item{z}{Categorical CTDS data stream indicating which cell was moved to, where \code{z=directions+1} indicates no movement from the current cell. When \code{directions}=4, z=1 indicates a move to the left cell, z=2 right cell, z=3 upward cell, and z=4 downward cell. When \code{directions}=8, z=1 indicates top-left cell, z=2 middle-left cell, z=3 bottom-left cell, z=4 top-right cell, z=5 middle-right cell, z=6 bottom-right cell, z=7 top-middle cell, and z=8 bottom-middle cell.}
\item{tau}{Time difference (in \code{Time.unit}) between successive observations}
\item{cellCross}{Integer indexing move(s) across multiple cells within a time step (if any)}
\item{...}{Additional data streams (if any)}
\item{...}{Covariates (if any). Covariates are needed for all neighboring cells and are indexed by \code{1,2,...,directions} using a suffix (e.g. \code{cov.1}, \code{cov.2}, \code{cov.3}, and \code{cov.4} for a covariate named 'cov' with \code{directions=4}). Additionally, a covariate named \code{noMove} is automatically included that indicates when no cell transition occurred (hence no transition out of the current state); the \code{noMove} covariate can be used as a state transition probability covariate to forbid transitions from the current state when there is no cell movement.}
}
\description{
This wrapper function for \code{\link[ctmcmove]{path2ctmc}} and \code{\link[ctmcmove]{ctmc2glm}} converts a \code{data.frame} of coordinates, other data streams, and non-spatial covariates to a \code{\link{momentuHMMData}} object that can be passed directly to \code{\link{fitCTHMM}} (or as a list to \code{\link{MIfitCTHMM}}).
}
\details{
\itemize{
\item Any moves to adjacent cells (as defined by \code{directions}; see \code{\link[raster]{adjacent}}) between times t and t+1 are assumed to occur at time t+1 (where \code{tau} = t+1 - t), regardless of the length of the time step (\code{tau}) and where in the current cell or in the neighboring cell that the corresponding (continuous-space) locations reside. 
This is done to facilitate the temporal alignment of the discrete-space movement process with other data streams and spatio-temporal covariates. The appropriateness of this simplifying assumption will depend on \code{tau}, the scale of movement, and the raster resolution.

\item If there are move(s) across multiple cells within a time step, data stream(s) other than \code{z} (the categorical CTDS data stream) are set to \code{NA} for these time step(s). 
These \code{NA} data stream values must be manually set based on the time spent in each cell if they are to be included in subsequent analysis. 
Time-varying spatial covariates for these time step(s) are set based on the z-value for the initial cell and, if different, these must also be set manually based on z-value(s) at the time cell(s) were crossed.
All such instances are indicated wherever the 'cellCross' field is > 0.

\item If \code{data} is a \code{\link{crwData}} object, the \code{\link{momentuHMMData}} object created by \code{prepCTDS} is based on the best predicted 
locations (i.e., \code{crwData$crwPredict$mu.x} and \code{crwData$crwPredict$mu.y}). Prior to using \code{prepCTDS}, additional data streams or covariates unrelated to location (including z-values associated with
\code{spatialCovs} and \code{spatialCovs.grad} raster stacks or bricks) can be merged with the \code{crwData} object using \code{\link{crawlMerge}}.
}
}
\references{
Hanks E. M., Hooten M. B., and Alldredge M. W. 2015. Continuous-time Discrete-space Models for Animal Movement. The Annals of Applied Statistics 9:145-165
}
