% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\title{Preprocessing of the data streams, including calculation of step length, turning angle, and covariates from location data}
\usage{
prepData(Data, type = c("LL", "UTM"), coordNames = NULL, covNames = NULL,
  spatialCovs = NULL)
}
\arguments{
\item{Data}{A dataframe of data streams, including optionally a field \code{ID}
(identifiers for the observed individuals), coordinates from which step length ('step') 
and turning angle ('angle') area calculated, and any covariates (with names matching \code{covNames}). 
If step length and turning angle are to be calculated from coordinates, the \code{coordNames} argument 
must identify the names for the longitunal and latitudinal coordinates (e.g., \code{coordNames=c("x","y")}).
With the exception of \code{ID} and coordinates, all variables in Data are treated as data streams unless identified
as covariates in \code{covNames}.}

\item{type}{\code{'LL'} if longitude/latitude provided (default), \code{'UTM'} if easting/northing.}

\item{coordNames}{Names of the columns of coordinates in the \code{Data} data frame. Default: \code{c("x","y")}.}

\item{covNames}{Names of any covariates in \code{Data} dataframe. Any variables in \code{Data} (other than \code{ID}) that are not identified in \code{covNames} are treated as data streams.}

\item{spatialCovs}{List of raster layer(s) for any spatial covariates not included in \code{Data}. Covariates specified by 'spatialCovs' are
extracted based on location data for each time step.}
}
\value{
An object \code{momentuHMMData}
}
\description{
Preprocessing of the data streams, including calculation of step length, turning angle, and covariates from location data
}
\examples{
coord1 <- c(1,2,3,4,5,6,7,8,9,10)
coord2 <- c(1,1,1,2,2,2,1,1,1,2)
Data <- data.frame(coord1=coord1,coord2=coord2)
d <- prepData(Data,type='UTM',coordNames=c("coord1","coord2"))

}
