% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCTDS.R
\name{simCTDS}
\alias{simCTDS}
\title{Simulation tool}
\usage{
simCTDS(
  nbAnimals = 1,
  nbStates = 2,
  dist,
  Par,
  beta = NULL,
  delta = NULL,
  formula = ~1,
  formulaDelta = NULL,
  mixtures = 1,
  formulaPi = NULL,
  covs = NULL,
  nbCovs = 0,
  rast,
  spatialCovs = NULL,
  spatialCovs.grad = NULL,
  directions = 4,
  normalize.gradients = FALSE,
  grad.point.decreasing = FALSE,
  zero.idx = integer(),
  moveState = FALSE,
  obsPerAnimal = c(500, 1500),
  initialPosition = c(0, 0),
  DM = NULL,
  userBounds = NULL,
  workBounds = NULL,
  betaRef = NULL,
  stateNames = NULL,
  model = NULL,
  matchModelObs = TRUE,
  states = FALSE,
  lambda = 1,
  errorEllipse = NULL,
  ncores = 1,
  export = NULL,
  keepSwitch = FALSE,
  kappa = NULL
)
}
\arguments{
\item{nbAnimals}{Number of observed individuals to simulate.}

\item{nbStates}{Number of behavioural states to simulate.}

\item{dist}{A named list indicating the probability distributions of the data streams. A data stream named \code{z} must be included, and it must be assigned the \code{"ctds"} distribution.}

\item{Par}{A named list containing vectors of initial state-dependent probability distribution parameters for 
each data stream specified in \code{dist}. The parameters should be in the order expected by the pdfs of \code{dist}. 

If \code{DM} is not specified for a given data stream, then \code{Par} 
is on the natural (i.e., real) scale of the parameters. However, if \code{DM} is specified for a given data stream, then 
\code{Par} must be on the working (i.e., beta) scale of the parameters, and the length of \code{Par} must match the number 
of columns in the design matrix. See details below.}

\item{beta}{Matrix of regression parameters for the transition rates.}

\item{delta}{Initial value for the initial distribution of the HMM. Default: \code{rep(1/nbStates,nbStates)}. If \code{formulaDelta} includes a formula, then \code{delta} must be specified
as a k x (\code{nbStates}-1) matrix, where k is the number of covariates and the columns correspond to states 2:\code{nbStates}.}

\item{formula}{Regression formula for the transition rate covariates. Default: \code{~1} (no covariate effect). In addition to allowing standard functions in R formulas
(e.g., \code{cos(cov)}, \code{cov1*cov2}, \code{I(cov^2)}), special functions include \code{cosinor(cov,period)} for modeling cyclical patterns, spline functions 
(\code{\link[splines]{bs}}, \code{\link[splines]{ns}}, \code{\link[splines2]{bSpline}}, \code{\link[splines2]{cSpline}}, \code{\link[splines2]{iSpline}}, and \code{\link[splines2]{mSpline}}), 
and state- or parameter-specific formulas (see details).
Any formula terms that are not state- or parameter-specific are included on all of the transition rates.}

\item{formulaDelta}{Regression formula for the initial distribution. Default: \code{NULL} (no covariate effects and \code{delta} is specified on the real scale). Standard functions in R formulas are allowed (e.g., \code{cos(cov)}, \code{cov1*cov2}, \code{I(cov^2)}). When any formula is provided, then \code{delta} must be specified on the working scale.}

\item{mixtures}{Number of mixtures for the state transition probabilities  (i.e. discrete random effects *sensu* DeRuiter et al. 2017). Default: \code{mixtures=1}.}

\item{formulaPi}{Regression formula for the mixture distribution probabilities. Default: \code{NULL} (no covariate effects; both \code{beta$pi} and \code{fixPar$pi} are specified on the real scale). Standard functions in R formulas are allowed (e.g., \code{cos(cov)}, \code{cov1*cov2}, \code{I(cov^2)}). When any formula is provided, then both \code{beta$pi} and \code{fixPar$pi} are specified on the working scale.
Note that only the covariate values corresponding to the first time step for each individual ID are used (i.e. time-varying covariates cannot be used for the mixture probabilties).}

\item{covs}{Covariate values to include in the simulated data, as a dataframe. The names of any covariates specified by \code{covs} can
be included in \code{formula} and/or \code{DM}. Covariates can also be simulated according to a standard normal distribution, by setting
\code{covs} to \code{NULL} (the default), and specifying \code{nbCovs>0}.}

\item{nbCovs}{Number of covariates to simulate (0 by default). Does not need to be specified if
\code{covs} is specified. Simulated covariates are provided generic names (e.g., 'cov1' and 'cov2' for \code{nbCovs=2}) and can be included in \code{formula} and/or \code{DM}.}

\item{rast}{A raster object or raster stack object that will define the discrete-space grid cells for the CTMC movement path. \code{spatialCovs} and \code{spatialCovs.grad} must have the same extent, number of rows and columns, projection, resolution, and origin as \code{rast}.}

\item{spatialCovs}{List of \code{\link[raster]{raster}} objects for spatio-temporally referenced covariates. Covariates specified by \code{spatialCovs} are extracted from the raster 
layer(s) based on the location data (and the z values for a raster \code{\link[raster]{stack}} 
or \code{\link[raster]{brick}}) for each time step.  If an element of \code{spatialCovs} is a raster \code{\link[raster]{stack}} or \code{\link[raster]{brick}}, 
then z values must be set using \code{raster::setZ} and \code{data} must include column(s) of the corresponding z value(s) for each observation (e.g., 'Date'). In the \code{\link{momentuHMMData}} object returned by \code{prepCTDS}, covariates for the current position (e.g.\ for use in \code{formula} or \code{DM}) are named with a \code{.cur} suffix (e.g. \code{cov1.cur}).}

\item{spatialCovs.grad}{List of \code{\link[raster]{raster}} objects for spatio-temporally referenced covariates, where a directional gradient is to be calculated internally using \code{\link[ctmcmove]{rast.grad}}. Gradient-based covariates specified by \code{spatialCovs.grad} are extracted from the raster 
layer(s) based on the location data (and the z values for a raster \code{\link[raster]{stack}} or \code{\link[raster]{brick}}) for each time step.  If an element of \code{spatialCovs.grad} is a raster \code{\link[raster]{stack}} or \code{\link[raster]{brick}}, 
then z values must be set using \code{raster::setZ} and \code{data} must include column(s) of the corresponding z value(s) for each observation (e.g., 'Date').}

\item{directions}{Integer. Either 4 (indicating a "Rook's neighborhood" of 4 neighboring grid cells) or 8 (indicating a "King's neighborhood" of 8 neighboring grid cells).}

\item{normalize.gradients}{Logical. Default is FALSE. If TRUE, then all gradient covariates for \code{spatialCovs.grad} are normalized by dividing by the length of the gradient vector at each point.}

\item{grad.point.decreasing}{Logical. If TRUE, then the gradient covariates are positive in the direction of decreasing values of the covariate. If FALSE, then the gradient covariates are positive in the direction of increasing values of the covariate (like a true gradient).}

\item{zero.idx}{Integer vector of the indices of raster cells that are not passable and should be excluded. These are cells where movement should be impossible. Default is zero.idx=integer().}

\item{moveState}{Logical indicating whether or not transitions out of the current state are forbidden when the animal does not move out of the current cell. Default: \code{FALSE}. Note that if \code{moveState} is \code{TRUE}, then the realized state transition rates will not accurately reflect \code{beta}.}

\item{obsPerAnimal}{Either the number of observations per animal (if single value) or the bounds of the number of observations per animal (if vector of two values). In the latter case, 
the numbers of obervations generated for each animal are uniformously picked from this interval. Alternatively, \code{obsPerAnimal} can be specified as
a list of length \code{nbAnimals} with each element providing the number of observations (if single value) or the bounds (if vector of two values) for each individual.
Default: \code{c(500,1500)}.}

\item{initialPosition}{2-vector providing the x- and y-coordinates of the initial position for all animals. Alternatively, \code{initialPosition} can be specified as
a list of length \code{nbAnimals} with each element a 2-vector providing the x- and y-coordinates of the initial position for each individual.
Default: \code{c(0,0)}.  If \code{mvnCoord} corresponds to a data stream with ``mvnorm3'' or ''rw_mvnorm3'' probability distributions, then \code{initialPosition} must be composed of 3-vector(s) for the x-, y-, and z-coordinates.}

\item{DM}{An optional named list indicating the design matrices to be used for the probability distribution parameters of each data 
stream. Each element of \code{DM} can either be a named list of regression formulas or a ``pseudo'' design matrix.}

\item{userBounds}{An optional named list of 2-column matrices specifying bounds on the natural (i.e, real) scale of the probability 
distribution parameters for each data stream.}

\item{workBounds}{An optional named list of 2-column matrices specifying bounds on the working scale of the probability distribution, transition probability, and initial distribution parameters. For each matrix, the first column pertains to the lower bound and the second column the upper bound.
For data streams, each element of \code{workBounds} should be a k x 2 matrix with the same name of the corresponding element of 
\code{Par}, where k is the number of parameters. For transition rate parameters, the corresponding element of \code{workBounds} must be a k x 2 matrix named ``beta'', where k=\code{length(beta)}. For initial distribution parameters, the corresponding element of \code{workBounds} must be a k x 2 matrix named ``delta'', where k=\code{length(delta)}.
\code{workBounds} is ignored for any given data stream unless \code{DM} is also specified.}

\item{betaRef}{Numeric vector of length \code{nbStates} indicating the reference elements for the state transition rate matrix. Default: NULL, in which case
the diagonal elements of the transition rate matrix are the reference.}

\item{stateNames}{Optional character vector of length nbStates indicating state names.}

\item{model}{A \code{\link{momentuHMM}}, \code{\link{miHMM}}, or \code{\link{miSum}} object. This option can be used to simulate from a fitted model.  Default: NULL.
Note that, if this argument is specified, most other arguments will be ignored -- except for \code{nbAnimals},
\code{obsPerAnimal}, \code{states}, \code{initialPosition}, \code{lambda}, \code{errorEllipse}, and, if covariate values different from those in the data should be specified, 
\code{covs}, and \code{spatialCovs}. It is not appropriate to simulate movement data from a \code{model} that was fitted to latitude/longitude data (because \code{simData} assumes Cartesian coordinates).}

\item{matchModelObs}{If \code{model} is provided, logical indicating whether to match \code{nbAnimals}, \code{obsPerAnimal}, and observation times to the fitted model data. If \code{TRUE}, then \code{nbAnimals}, \code{obsPerAnimal}, and \code{lambda} are ignored. Default: \code{TRUE}.}

\item{states}{\code{TRUE} if the simulated states should be returned, \code{FALSE} otherwise (default).}

\item{lambda}{Observation rate. \code{lambda} is the rate parameter of the exponential distribution for the waiting times between successive observations, i.e., 
\code{1/lambda} is the expected time between successive location observations. If \code{model} is specified and \code{model$data} time column is of class \code{\link[base]{date-time}} or \code{\link[base]{date}}, \code{lambda} has the same units as the \code{Time.unit} argument in \code{\link{prepCTDS}}. Default: 1.}

\item{errorEllipse}{List providing the upper bound for the semi-major axis (\code{M}; on scale of x- and y-coordinates), semi-minor axis (\code{m}; 
on scale of x- and y-coordinates), and orientation (\code{r}; in degrees) of location error ellipses. If \code{NULL} (the default), no location 
measurement error is simulated. If \code{errorEllipse} is specified, then each observed location is subject to bivariate normal errors as described 
in McClintock et al. (2015), where the components of the error ellipse for each location are randomly drawn from \code{runif(1,min(errorEllipse$M),max(errorEllipse$M))}, 
\code{runif(1,min(errorEllipse$m),max(errorEllipse$m))}, and \code{runif(1,min(errorEllipse$r),max(errorEllipse$r))}. If only a single value is provided for any of the 
error ellipse elements, then the corresponding component is fixed to this value for each location. Only coordinate data streams are subject to location measurement error;
any other data streams are observed without error.}

\item{ncores}{Number of cores to use for parallel processing. Default: 1 (no parallel processing).}

\item{export}{Character vector of the names of any additional objects or functions in the global environment that are used in \code{DM}, \code{formula}, \code{formulaDelta}, and/or \code{formulaPi}. Only necessary if \code{ncores>1} so that the needed items will be exported to the workers.}

\item{keepSwitch}{Logical indicating whether or not to return the (typically unobserved) data at the times when potential state switches could have occurred. Default: \code{FALSE}. If set to \code{TRUE}, an additional logical field named \code{isObs} is returned, where \code{TRUE} indicates observations and \code{FALSE} indicates state switches.}

\item{kappa}{List of the form \code{list(method=c("all","random","quantile"),nspCov=NA,spCov=NA)} defining the method for obtaining the upper bound for the transition rate out of the current state (see Blackwell et al. 2016). The list can include up to three named objects: 1) \code{method}, a character string indicating the method for calculating the upper bound based on the covariates in the model (\code{"all"}, \code{"random"}, or \code{"quantile"}); 2) \code{nspCov}, a positive scalar for subsampling the non-spatial covariates (when \code{method="random"} or \code{method="quantile"}); and \code{spCov}, a positive scalar for subsampling the spatial covariates (when \code{method="random"} or \code{method="quantile"}).
Default method is \code{"all"}, in which case \code{kappa} is calculated based on all of the observed covariate values (note this can be slow and/or memory could become an issue for large datasets and/or rasters). For \code{method="random"}, the observed covariates are subsampled with up to \code{nspCov} samples of any non-spatial covariates and up to \code{spCov} samples of any spatial covariates (defaults are \code{1000} for \code{nspCov} and \code{10000} for \code{spCov}). For \code{method="quantile"}, all combinations of 100-length sequences spanning the (\code{nspCov}/2, 1-\code{nspCov}/2) and (\code{spCov}/2, 1-\code{spCov}/2) quantiles of the covariates are used (defaults are \code{0.05} for both \code{nspCov} and \code{spCov}). 
Ignored unless covariates are included in \code{formula}.
 
Alternatively, \code{kappa} can be manually specified as a finite positive scalar (instead of a list) indicating the maximum allowed value for the row sums of the off-diagonal elements in the state transition rate matrix, such that the minimum value for the diagonal elements is \code{-kappa}. In this case, the transition rate parameters (\code{beta}) are specified on the logit scale (instead of the log scale).
If \code{model} is provided and \code{kappa} is \code{NULL}, then \code{kappa} is set to \code{model$conditions$kappa} (see \code{\link{fitCTHMM}}).}
}
\value{
If the simulated data have no measurement error (i.e., \code{errorEllipse=NULL}), a \code{\link{momentuHMMData}} object, 
i.e., a dataframe of:
\item{ID}{The ID(s) of the observed animal(s)}
\item{time}{Numeric time of each observed observation}
\item{...}{Data streams as specified by \code{dist}}
\item{x}{Easting location cell centroid}
\item{y}{Northing location cell centroid}
\item{...}{Covariates (if any). If \code{moveState=TRUE}, then a covariate named \code{noMove} is included that indicates when no cell transition occurred (hence no transition out of the current state); the \code{noMove} covariate can then be used as a state transition probability covariate to forbid transitions from the current state when there is no cell movement.}

If simulated location data include measurement error (i.e., \code{errorEllipse!=NULL}), a dataframe of:
\item{ID}{The ID(s) of the observed animal(s)}
\item{time}{Numeric time of each observed (and missing) observation}
\item{z}{Categorical index indicating observed cell movement (where \code{z=(directions+1)} indicates no movement)}
\item{x}{Easting observed location cell centroid}
\item{y}{Norting observed location cell centroid}
\item{tau}{Time difference between consecutive observations}
\item{...}{Data streams that are not derived from location (if applicable)}
\item{...}{Covariates at true (\code{mux},\code{muy}) locations (if any) and neighboring cell locations (with suffixes indicating neighbor, e.g., \code{cov.1}, \code{cov.2}, ..., \code{cov.directions}). If \code{moveState=TRUE}, then a covariate named \code{paste0("z",(directions+1))} is included that indicates when no cell transition occurred (hence no transition out of the current state).}
\item{mux}{Easting true location cell centroid}
\item{muy}{Northingtrue location cell centroid}
\item{error_semimajor_axis}{error ellipse semi-major axis (if applicable)}
\item{error_semiminor_axis}{error ellipse semi-minor axis (if applicable)}
\item{error_ellipse_orientation}{error ellipse orientation (if applicable)}
\item{ln.sd.x}{log of the square root of the x-variance of bivariate normal error (if applicable; required for error ellipse models in \code{\link{crawlWrap}})}
\item{ln.sd.y}{log of the square root of the y-variance of bivariate normal error (if applicable; required for error ellipse models in \code{\link{crawlWrap}})}
\item{error.corr}{correlation term of bivariate normal error (if applicable; required for error ellipse models in \code{\link{crawlWrap}})}
}
\description{
Simulates data from a (multivariate) continuous-time discrete-space hidden Markov model based on Blackwell et al. (2016), where specification of \code{kappa} determines the potential times of state switches when \code{formula} includes covariates (larger \code{kappa} means potential switches are more frequent). Note that state transitions can only occur when there is a cell transition (unless \code{moveState} is set to \code{FALSE}), and time-varying covariates are assumed piece-wise constant between observations. Movement data are assumed to be in Cartesian coordinates (not longitude/latitude) and can be generated with or without observation error attributable to location measurement error.
}
\details{
\itemize{
\item \code{simCTDS} assumes the snapshot property applies to all data stream distributions (i.e. observations are "instantaneous") except for the continuous-time discrete-space (\code{ctds}), (multivariate) normal random walk (\code{rw_norm}, \code{rw_mvnorm2}, \code{rw_mvnorm3}), and Poisson (\code{pois}) distributions. For these particular distributions, the observed data are not "instantaneous"; they depend on the time interval between observations \eqn{(\Delta_t)} and, hence, the state sequence during the entire interval.
If fitting with \code{\link{fitCTHMM}} (or \code{\link{MIfitCTHMM}}), it is critical that the frequency of observations (specified by \code{lambda}) is high relative to the serial correlation in the hidden state process (specified by \code{beta}) in order for the discrete-time approximation of \code{\link{fitCTHMM}} to be reasonably accurate for these distributions.

\item If the length of covariate values passed (either through 'covs', or 'model') is not the same
as the number of observations suggested by 'nbAnimals' and 'obsPerAnimal', then the series of
covariates is either shortened (removing last values - if too long) or extended (starting
over from the first values - if too short).

\item When covariates are not included in \code{formulaDelta} (i.e. \code{formulaDelta=NULL}), then \code{delta} is specified as a vector of length \code{nbStates} that 
sums to 1.  When covariates are included in \code{formulaDelta}, then \code{delta} must be specified
as a k x (\code{nbStates}-1) matrix of working parameters, where k is the number of regression coefficients and the columns correspond to states 2:\code{nbStates}. For example, in a 3-state
HMM with \code{formulaDelta=~cov1+cov2}, the matrix \code{delta} has three rows (intercept + two covariates)
and 2 columns (corresponding to states 2 and 3). The initial distribution working parameters are transformed to the real scale as \code{exp(covsDelta*Delta)/rowSums(exp(covsDelta*Delta))}, where \code{covsDelta} is the N x k design matrix, \code{Delta=cbind(rep(0,k),delta)} is a k x \code{nbStates} matrix of working parameters,
and \code{N=length(unique(data$ID))}.
}
}
