% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trMatrix_rcpp}
\alias{trMatrix_rcpp}
\title{Transition probability matrix}
\usage{
trMatrix_rcpp(
  nbStates,
  beta,
  covs,
  betaRef,
  CT = FALSE,
  dt = as.numeric(c()),
  aInd = as.integer(c()),
  rateMatrix = FALSE,
  kappa = NA_real_,
  check = TRUE
)
}
\arguments{
\item{nbStates}{Number of states}

\item{beta}{Matrix of regression parameters}

\item{covs}{Matrix of covariate values}

\item{betaRef}{Indices of reference elements for t.p.m. multinomial logit link.}

\item{CT}{logical indicating discrete-time approximation of a continuous-time model}

\item{dt}{numeric vector of length \code{nrow(covs)} indicating the time difference between observations. Ignored unless \code{CT=TRUE}.}

\item{aInd}{Vector of indices of the rows at which the data (i.e. \code{covs}) switches to another animal. Ignored unless \code{CT=TRUE}.}

\item{rateMatrix}{logical indicating whether or not to return the transition rate matrix. Ignored unless \code{CT=TRUE}.}

\item{kappa}{maximum allowed value for the row sums of the off-diagonal elements in the state transition rate matrix, such that the minimum value for the diagonal elements is \code{-kappa}. Default: \code{Inf}. Setting less than \code{Inf} can help avoid numerical issues during optimization, in which case the transition rate parameters \code{beta} are on the logit scale (instead of the log scale).}

\item{check}{logical indicating whether or not to check transition probability matrix for issues. Ignored unless \code{CT=TRUE}.}
}
\value{
Three dimensional array \code{trMat}, such that \code{trMat[,,t]} is the transition matrix at
time t.
}
\description{
Computation of the transition probability matrix, as a function of the covariates and the regression
parameters. Written in C++. Used in \code{\link{viterbi}}.
}
