% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trMatrix_rcpp}
\alias{trMatrix_rcpp}
\title{Transition probability matrix}
\usage{
trMatrix_rcpp(nbStates, beta, covs)
}
\arguments{
\item{nbStates}{Number of states}

\item{beta}{Matrix of regression parameters}

\item{covs}{Matrix of covariate values}
}
\value{
Three dimensional array \code{trMat}, such that \code{trMat[,,t]} is the transition matrix at
time t.
}
\description{
Computation of the transition probability matrix, as a function of the covariates and the regression
parameters. Written in C++. Used in \code{\link{viterbi}}.
}
