% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnAngle.R
\name{turnAngle}
\alias{turnAngle}
\title{Turning angle}
\usage{
turnAngle(x, y, z)
}
\arguments{
\item{x}{First point}

\item{y}{Second point}

\item{z}{Third point}
}
\value{
The angle between vectors (x,y) and (y,z)
}
\description{
Used in \code{\link{prepData}} and \code{\link{simData}}.
}
\examples{
\dontrun{
x <- c(0,0)
y <- c(4,6)
z <- c(10,7)
momentuHMM:::turnAngle(x,y,z)
}
}
