% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{XBloop_rcpp}
\alias{XBloop_rcpp}
\title{Get XB}
\usage{
XBloop_rcpp(
  DM,
  Xvec,
  nbObs,
  nr,
  nc,
  circularAngleMean,
  consensus,
  rindex,
  cindex,
  nbStates,
  refCoeff = 1
)
}
\arguments{
\item{DM}{design matrix}

\item{Xvec}{working parameters}

\item{nbObs}{number of observations}

\item{nr}{number of rows in design matrix}

\item{nc}{number of column in design matrix}

\item{circularAngleMean}{indicator for whether or not circular-circular regression model}

\item{consensus}{indicator for whether or not circular-circular regression consensus model}

\item{rindex}{row index for design matrix}

\item{cindex}{column index for design matrix}

\item{nbStates}{number of states}

\item{refCoeff}{intercept coefficient for circular-circular regression model}
}
\value{
XB matrix
}
\description{
Loop for computation of design matrix (X) times the working scale parameters (B). Written in C++. Used in \code{\link{w2n}}.
}
